Use ls-files by default.
Not-quite-empirical testing has shown that reading .gitignore files is actually
negligibly slower than using 'git ls-files', and since if I were to implement
negated ignores for .gitignore-file use the end result would be essentially the
same as using 'ls-files', let's just use it by default.

The one VERY MINOR advantage that the wildcards, etc. from .gitignore files has
over using ls-files, is that you wouldn't need to re-enter a buffer before a
newly created file would start to be ignored (because it would be matched by
the existing ignores). But this advantage is very small, since if you're in
Vim, and you create a new file, you most likely will end up switching to that
buffer to edit it, and so when you switch back to the original buffer, the
ignore list will be updated then, and that new file is ignored as you would
expect!

tl;dr No reason not to use ls-files by default.